refactor circuit breaker test
